// gauge-ts/src/gen/messages.ts
export interface KillProcessRequest {
}

/**
 * / Sends to any request which needs a execution status as response
 * / usually step execution, hooks etc will return this
 */
export interface ExecutionStatusResponse {
  /** / Holds the suite result after suite execution done. */
  executionResult: ProtoExecutionResult | undefined;
}
// ...
export interface ExecutionEndingRequest {
  /** / Holds the current suite execution info. */
  currentExecutionInfo:
    | ExecutionInfo
    | undefined;
  /** / Holds the suite result in execution ending. */
  suiteResult: ProtoSuiteResult | undefined;
  stream: number;
}
// ...
export interface ExecuteStepRequest {
  /**
   * / Contains the actual text of the Step being executed.
   * / This contains the parameters as defined in the Spec.
   */
  actualStepText: string;
  /**
   * / Contains the parsed text of the Step being executed.
   * / The paramters are replaced with placeholders.
   */
  parsedStepText: string;
  /** / Flag to indicate if the execution of the Scenario, containing the current Step, failed. */
  scenarioFailing: boolean;
  /** / Collection of parameters applicable to the current Step. */
  parameters: Parameter[];
  stream: number;
}
// ...
export interface CacheFileRequest {
  /** / File content of the file to be cached */
  content: string;
  /** / File path of the file to be cached */
  filePath: string;
  /** / Specifies if the file is closed */
  isClosed: boolean;
  /** / Specifies the status of the file */
  status: CacheFileRequest_FileStatus;
}

export enum CacheFileRequest_FileStatus {
  /** CHANGED - / The file content was changed in the client */
  CHANGED = 0,
  /** CLOSED - / The file was closed in the client */
  CLOSED = 1,
  /** CREATED - / The file was created on the client */
  CREATED = 2,
  /** DELETED - / The file was deleted on the client */
  DELETED = 3,
  /** OPENED - / The file is opened in the client */
  OPENED = 4,
  UNRECOGNIZED = -1,
}
// ...
export interface ImplementationFileGlobPatternResponse {
  /** / List of implementation file glob patterns */
  globPatterns: string[];
}
// ...
export interface Empty {
}
// ...
export const ExecutionStatusResponse: MessageFns<ExecutionStatusResponse> = {
  encode(message: ExecutionStatusResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.executionResult !== undefined) {
      ProtoExecutionResult.encode(message.executionResult, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ExecutionStatusResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseExecutionStatusResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.executionResult = ProtoExecutionResult.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ExecutionStatusResponse {
    return {
      executionResult: isSet(object.executionResult)
        ? ProtoExecutionResult.fromJSON(object.executionResult)
        : undefined,
    };
  },

  toJSON(message: ExecutionStatusResponse): unknown {
    const obj: any = {};
    if (message.executionResult !== undefined) {
      obj.executionResult = ProtoExecutionResult.toJSON(message.executionResult);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ExecutionStatusResponse>, I>>(base?: I): ExecutionStatusResponse {
    return ExecutionStatusResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ExecutionStatusResponse>, I>>(object: I): ExecutionStatusResponse {
    const message = createBaseExecutionStatusResponse();
    message.executionResult = (object.executionResult !== undefined && object.executionResult !== null)
      ? ProtoExecutionResult.fromPartial(object.executionResult)
      : undefined;
    return message;
  },
};
// ...
export const ExecuteStepRequest: MessageFns<ExecuteStepRequest> = {
  encode(message: ExecuteStepRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.actualStepText !== "") {
      writer.uint32(10).string(message.actualStepText);
    }
    if (message.parsedStepText !== "") {
      writer.uint32(18).string(message.parsedStepText);
    }
    if (message.scenarioFailing !== false) {
      writer.uint32(24).bool(message.scenarioFailing);
    }
    for (const v of message.parameters) {
      Parameter.encode(v!, writer.uint32(34).fork()).join();
    }
    if (message.stream !== 0) {
      writer.uint32(40).int32(message.stream);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ExecuteStepRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseExecuteStepRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.actualStepText = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.parsedStepText = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.scenarioFailing = reader.bool();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.parameters.push(Parameter.decode(reader, reader.uint32()));
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.stream = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ExecuteStepRequest {
    return {
      actualStepText: isSet(object.actualStepText) ? globalThis.String(object.actualStepText) : "",
      parsedStepText: isSet(object.parsedStepText) ? globalThis.String(object.parsedStepText) : "",
      scenarioFailing: isSet(object.scenarioFailing) ? globalThis.Boolean(object.scenarioFailing) : false,
      parameters: globalThis.Array.isArray(object?.parameters)
        ? object.parameters.map((e: any) => Parameter.fromJSON(e))
        : [],
      stream: isSet(object.stream) ? globalThis.Number(object.stream) : 0,
    };
  },

  toJSON(message: ExecuteStepRequest): unknown {
    const obj: any = {};
    if (message.actualStepText !== "") {
      obj.actualStepText = message.actualStepText;
    }
    if (message.parsedStepText !== "") {
      obj.parsedStepText = message.parsedStepText;
    }
    if (message.scenarioFailing !== false) {
      obj.scenarioFailing = message.scenarioFailing;
    }
    if (message.parameters?.length) {
      obj.parameters = message.parameters.map((e) => Parameter.toJSON(e));
    }
    if (message.stream !== 0) {
      obj.stream = Math.round(message.stream);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ExecuteStepRequest>, I>>(base?: I): ExecuteStepRequest {
    return ExecuteStepRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ExecuteStepRequest>, I>>(object: I): ExecuteStepRequest {
    const message = createBaseExecuteStepRequest();
    message.actualStepText = object.actualStepText ?? "";
    message.parsedStepText = object.parsedStepText ?? "";
    message.scenarioFailing = object.scenarioFailing ?? false;
    message.parameters = object.parameters?.map((e) => Parameter.fromPartial(e)) || [];
    message.stream = object.stream ?? 0;
    return message;
  },
};
// ...
export const CacheFileRequest: MessageFns<CacheFileRequest> = {
  encode(message: CacheFileRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.content !== "") {
      writer.uint32(10).string(message.content);
    }
    if (message.filePath !== "") {
      writer.uint32(18).string(message.filePath);
    }
    if (message.isClosed !== false) {
      writer.uint32(24).bool(message.isClosed);
    }
    if (message.status !== 0) {
      writer.uint32(32).int32(message.status);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CacheFileRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCacheFileRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.content = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.filePath = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.isClosed = reader.bool();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.status = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CacheFileRequest {
    return {
      content: isSet(object.content) ? globalThis.String(object.content) : "",
      filePath: isSet(object.filePath) ? globalThis.String(object.filePath) : "",
      isClosed: isSet(object.isClosed) ? globalThis.Boolean(object.isClosed) : false,
      status: isSet(object.status) ? cacheFileRequest_FileStatusFromJSON(object.status) : 0,
    };
  },

  toJSON(message: CacheFileRequest): unknown {
    const obj: any = {};
    if (message.content !== "") {
      obj.content = message.content;
    }
    if (message.filePath !== "") {
      obj.filePath = message.filePath;
    }
    if (message.isClosed !== false) {
      obj.isClosed = message.isClosed;
    }
    if (message.status !== 0) {
      obj.status = cacheFileRequest_FileStatusToJSON(message.status);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CacheFileRequest>, I>>(base?: I): CacheFileRequest {
    return CacheFileRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CacheFileRequest>, I>>(object: I): CacheFileRequest {
    const message = createBaseCacheFileRequest();
    message.content = object.content ?? "";
    message.filePath = object.filePath ?? "";
    message.isClosed = object.isClosed ?? false;
    message.status = object.status ?? 0;
    return message;
  },
};
// ...
export const Empty: MessageFns<Empty> = {
  encode(_: Empty, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Empty {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseEmpty();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(_: any): Empty {
    return {};
  },

  toJSON(_: Empty): unknown {
    const obj: any = {};
    return obj;
  },

  create<I extends Exact<DeepPartial<Empty>, I>>(base?: I): Empty {
    return Empty.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Empty>, I>>(_: I): Empty {
    const message = createBaseEmpty();
    return message;
  },
};