// gauge-ts/src/gen/services.ts
export type RunnerService = typeof RunnerService;
// ...
export const RunnerClient = makeGenericClientConstructor(RunnerService, "gauge.messages.Runner") as unknown as {
  new (address: string, credentials: ChannelCredentials, options?: Partial<ClientOptions>): RunnerClient;
  service: typeof RunnerService;
  serviceName: string;
};